+++ .gitignore
@@ -68,6 +68,7 @@ build/
 # Published via prepublishOnly; built before test via pretest. Grows as modules migrate.
 /tsconfig.build.tsbuildinfo
 /gsd-core/bin/lib/host-integration.cjs
+/gsd-core/bin/lib/install-engine.cjs
 /gsd-core/bin/lib/capability-loader.cjs
 /gsd-core/bin/lib/capability-source.cjs
 /gsd-core/bin/lib/capability-ledger.cjs
