+++ tests/bug-924-claude-flat-skill-layout.test.cjs
@@ -37,7 +37,7 @@ const os = require('node:os');
 const ROOT = path.join(__dirname, '..');
 const COMMANDS_GSD = path.join(ROOT, 'commands', 'gsd');
 
-const { installRuntimeArtifacts } = require('../bin/install.js');
+const { installRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
 const { cleanup } = require('./helpers.cjs');
 const {
   loadSkillsManifest,
