+++ tests/copilot-install.test.cjs
@@ -44,12 +44,13 @@ const {
   writeCopilotHookConfig,
   writeManifest,
   reportLocalPatches,
-  installRuntimeArtifacts,
   runtimeMap,
   allRuntimes,
   parseRuntimeInput,
   buildRuntimePromptText,
 } = require('../bin/install.js');
+
+const { installRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
 
 const { getGlobalConfigDir } = require('../gsd-core/bin/lib/runtime-homes.cjs');
