+++ tests/enh-790-augment-commands.test.cjs
@@ -21,7 +21,7 @@ const path = require('node:path');
 
 const { createTempDir, cleanup } = require('./helpers.cjs');
 
-const { installRuntimeArtifacts } = require('../bin/install.js');
+const { installRuntimeArtifacts, uninstallRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
 const { resolveRuntimeArtifactLayout } = require('../gsd-core/bin/lib/runtime-artifact-layout.cjs');
 const { loadSkillsManifest, resolveProfile } = require('../gsd-core/bin/lib/install-profiles.cjs');
 
@@ -177,7 +177,7 @@ describe('enh-790 — uninstallRuntimeArtifacts removes augment commands', () =>
     const configDir = createTempDir('gsd-enh790-uninstall-');
     t.after(() => cleanup(configDir));
 
-    const { uninstallRuntimeArtifacts } = require('../bin/install.js');
+    // uninstallRuntimeArtifacts is imported from install-engine.cjs at the top of this file
 
     // Pre-create: a GSD command + a user-owned command
     const commandsDir = path.join(configDir, 'commands');
