+++ tests/fixtures/golden-install-parity/copilot.json
@@ -96,6 +96,7 @@
   "gsd-core/bin/lib/host-integration.cjs": "5050b73b4c50fc79",
   "gsd-core/bin/lib/init-command-router.cjs": "5f6b583acaba1037",
   "gsd-core/bin/lib/init.cjs": "0ada8bfab0200aac",
+  "gsd-core/bin/lib/install-engine.cjs": "ebe23c56af4637de",
   "gsd-core/bin/lib/install-profiles.cjs": "8c5d1318f4940307",
   "gsd-core/bin/lib/installer-migration-authoring.cjs": "050285a34b4b0b22",
   "gsd-core/bin/lib/installer-migration-report.cjs": "da74c64aec42cf6d",
