+++ tests/hermes-skills-migration.test.cjs
@@ -21,8 +21,9 @@ const fs = require('fs');
 
 const {
   convertClaudeCommandToClaudeSkill,
-  installRuntimeArtifacts,
 } = require('../bin/install.js');
+
+const { installRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
 const { parseFrontmatter, cleanup } = require('./helpers.cjs');
 const pkg = require('../package.json');
